Move shadow_css into stylesheet_compiler/ subpackage.
PiperOrigin-RevId: 228807226
